+++ tests/Unit/Iso3166NumericTest.php
@@ -54,4 +54,14 @@ final class Iso3166NumericTest extends TestCase
         $rule->setErrorMessage($customErrorMessage);
         $rule->validate('attr', 'invalid', $closure(...));
     }
+
+    /** Regression test for bugfix, see tag 1.0.1 */
+    public function testSetErrorMessageReturnRuleInstance(): void
+    {
+        $customErrorMessage = 'Message';
+        $rule = new Iso3166Numeric();
+        $ruleInstance = $rule->setErrorMessage($customErrorMessage);
+
+        self::assertInstanceOf(Iso3166Numeric::class, $ruleInstance);
+    }
 }